refactor: Participants Validator 분리
